==> mr_clinic_model/6_models_evaluation.R <==
# load packages and root path
if (T) { 
    suppressPackageStartupMessages({
        library(survcomp)
        library(glmnet)
        library(survival)
        library(caret)
        library(randomForestSRC)
        library(ggplot2)
        library(pheatmap)
    })
    if (dir.exists("/wangshuo")) {
        root = "/wangshuo/zhaox"
    } else {
        root = "/home/tongxueqing/zhao"
    }
    root = file.path(root, "ImageProcessing/mr_clinic_model/_data")
    set.seed(1)

    to_ci = function(set, cox = cox.new) {
        pred = predict(cox, newdata = set, type = "lp")
        ci = concordance.index(x = pred, surv.time = set$time, surv.event = set$event, method = "noether")    
        ci$c.index
    }

    combine_pred = function(set, models = L) {
        preds = sapply(models, function(cox) {
            pred = predict(cox, newdata = set, type = "lp")
        })
        pred = rowMeans(preds)
        ci = concordance.index(x = pred, surv.time = set$time, surv.event = set$event, method = "noether")
        ci$c.index
    }
}

# load mr data
if (T) {
    data = read.csv(file.path(root, "mr.iccfiltered.csv"), header = T, row.names = 1, stringsAsFactors = F)
    labeldata = read.csv(file.path(root, "clinic/info.csv"), header = T, row.names = 1, stringsAsFactors = F)
    data[,c("time", "event")] = labeldata[match(data$name, labeldata$name), c('time', 'event')]
    common.name = intersect(data$name, labeldata$name)
    data = data[data$name %in% common.name,]
    data$time = labeldata$time[match(data$name, labeldata$name)]
    data$event = labeldata$event[match(data$name, labeldata$name)]

    features = colnames(data)[1:(ncol(data) - 7)]

    names = unique(data$name)
    newdata = data.frame(name = names)
    for (serie in unique(data$series)) {
        subdata = data[data$series == serie,]
        subdata = subdata[match(subdata$name, names), features]
        colnames(subdata) = paste(colnames(subdata), serie, sep = "_")
        for (col in colnames(subdata)) {
            newdata[,col] = subdata[,col]
        }
    }
    newdata$time = data[match(names, data$name), "time"]
    newdata$event = data[match(names, data$name), "event"]
    newdata$set = data[match(names, data$name), "set"]

    # k-fold prepare
    if (T) {
        valname = matrix(rep(0, 44 * 4), c(44, 4))
        v1 = c(1510261, 1706442, 1503005, 1614914, 1504357, 1510123, 1604846,
            1410116, 1511382, 1602296, 1501049, 1512396, 1503058, 1703032,
            1501050, 1407606, 1603591, 1602748, 1700572, 1509612, 1405541,
            1504545, 1401438, 1501684, 1404116, 1411030, 1409660, 1408401,
            1401150, 1500695, 1510358, 1507884, 1511384, 1600294, 1401942,
            1400974, 1407348, 1501417, 1700462, 1612055, 1613361, 1506707,
            1604426, 1602751)
        v2 = c(1610256, 1403311, 1404316, 1501611, 1702583, 1615309, 1605686,
            1401920, 1501820, 1404637, 1406199, 1405036, 1505902, 1701138,
            1508191, 1511058, 1601789, 1501010, 1615495, 1605918, 1502357,
            1604547, 1502675, 1608527, 1700618, 1409419, 1607972, 1404924,
            1601239, 1609874, 1613107, 1604176, 1604103, 1614786, 1505470,
            1701811, 1508384, 1601095, 1408018, 1702593, 1604060, 1510827,
            1512377, 1602586)
        v3 = c(1403602, 1402805, 1607174, 1509038, 1701668, 1508005, 1500694,
            1700457, 1403098, 1406407, 1601350, 1404315, 1607688, 1601181,
            1600296, 1404309, 1607236, 1406322, 1405481, 1602238, 1504186,
            1403257, 1508466, 1505057, 1503175, 1602926, 1702746, 1700342,
            1608980, 1402299, 1505414, 1402355, 1508163, 1501416, 1402909,
            1509570, 1505796, 1406489, 1701663, 1614913, 1507284, 1603444,
            1405152, 1509767)
        v4 = c(1605606, 1701475, 1610465, 1507248, 1608839, 1608463, 1512384,
            1401317, 1605283, 1508362, 1400860, 1407127, 1405035, 1700177,
            1505679, 1614456, 1510392, 1602186, 1504406, 1501415, 1605530,
            1702589, 1704471, 1507472, 1510442, 1410317, 1600991, 1602137,
            1503718, 1602620, 1615043, 1512101, 1605385, 1510941, 1403144,
            1507217, 1600870, 1602453, 1607437, 1510385, 1408272, 1410575,
            1510039, 1604044)
        valname[,1] = labeldata$name[match(v1, labeldata$number)]
        valname[,2] = labeldata$name[match(v2, labeldata$number)]
        valname[,3] = labeldata$name[match(v3, labeldata$number)]
        valname[,4] = labeldata$name[match(v4, labeldata$number)]
    }

    test = newdata[newdata$set == 1,]
}

# load clinic data
if (T) {
    data = read.csv(file.path(root, "clinic/ClinicMessageForAnalysis.csv"), stringsAsFactors = F)
    colnames(data) = c(
        "name", "realname", "smoke", "family.history", "number", 
        "T.read", "N.read", "total.cut", "gender", "age", "body.status", 
        "neuron", "EVB", "HB", "LDH", "sarcoma", "necrosis", "lymphocyte", 
        "N.cut.N3b", "total.cut.IVA", "event", "time"
        )
    data = data[,!grepl("(EVB|LDH|realname)", colnames(data))]
    data = na.omit(data)

    data$binary.smoke = ifelse(data$smoke > 300, 1, 0)
    data$smoke = NULL
    data$binary.body.status = ifelse(data$body.status > 85, 1, 0)
    data$body.status = NULL

    if (T) {
        valname = matrix(rep(0, 44 * 4), c(44, 4))
        v1 = c(1510261, 1706442, 1503005, 1614914, 1504357, 1510123, 1604846,
            1410116, 1511382, 1602296, 1501049, 1512396, 1503058, 1703032,
            1501050, 1407606, 1603591, 1602748, 1700572, 1509612, 1405541,
            1504545, 1401438, 1501684, 1404116, 1411030, 1409660, 1408401,
            1401150, 1500695, 1510358, 1507884, 1511384, 1600294, 1401942,
            1400974, 1407348, 1501417, 1700462, 1612055, 1613361, 1506707,
            1604426, 1602751)
        v2 = c(1610256, 1403311, 1404316, 1501611, 1702583, 1615309, 1605686,
            1401920, 1501820, 1404637, 1406199, 1405036, 1505902, 1701138,
            1508191, 1511058, 1601789, 1501010, 1615495, 1605918, 1502357,
            1604547, 1502675, 1608527, 1700618, 1409419, 1607972, 1404924,
            1601239, 1609874, 1613107, 1604176, 1604103, 1614786, 1505470,
            1701811, 1508384, 1601095, 1408018, 1702593, 1604060, 1510827,
            1512377, 1602586)
        v3 = c(1403602, 1402805, 1607174, 1509038, 1701668, 1508005, 1500694,
            1700457, 1403098, 1406407, 1601350, 1404315, 1607688, 1601181,
            1600296, 1404309, 1607236, 1406322, 1405481, 1602238, 1504186,
            1403257, 1508466, 1505057, 1503175, 1602926, 1702746, 1700342,
            1608980, 1402299, 1505414, 1402355, 1508163, 1501416, 1402909,
            1509570, 1505796, 1406489, 1701663, 1614913, 1507284, 1603444,
            1405152, 1509767)
        v4 = c(1605606, 1701475, 1610465, 1507248, 1608839, 1608463, 1512384,
            1401317, 1605283, 1508362, 1400860, 1407127, 1405035, 1700177,
            1505679, 1614456, 1510392, 1602186, 1504406, 1501415, 1605530,
            1702589, 1704471, 1507472, 1510442, 1410317, 1600991, 1602137,
            1503718, 1602620, 1615043, 1512101, 1605385, 1510941, 1403144,
            1507217, 1600870, 1602453, 1607437, 1510385, 1408272, 1410575,
            1510039, 1604044)
        valname[,1] = data$name[match(v1, data$number)]
        valname[,2] = data$name[match(v2, data$number)]
        valname[,3] = data$name[match(v3, data$number)]
        valname[,4] = data$name[match(v4, data$number)]
    }

    data$number = NULL
    data$set = 0
    # for (col in colnames(data)[!grepl("(name|set|event|time|HB|age)", colnames(data))]) {
    #     data[,col] = as.factor(data[,col])
    # }
    if (T) {
        data$binary.HB = ifelse(data$gender == 1, ifelse(data$HB < 130, 1, 0), ifelse(data$HB < 120, 1, 0))
        data$HB = NULL
    }
    newdata = data
    features = colnames(data)[!grepl("(name|set|event|time)", colnames(data))]
}

Ls = readRDS(file.path(root, "Ls_clinic_biHB.rds"))

results = t(sapply(1:length(Ls), function(i) {
    L = Ls[[i]]
    citrs = mean(sapply(1:4, function(k) {
        tr = newdata[newdata$set == 0 & !newdata$name %in% valname[,k],]
        to_ci(tr, L[[k]])
    }))
    civls = mean(sapply(1:4, function(k) {
        vl = newdata[newdata$set == 0 & newdata$name %in% valname[,k],]
        to_ci(vl, L[[k]])
    }))
    cits = combine_pred(newdata[newdata$set == 0,], L)
    c(citrs, civls, cits)
}))
L = Ls[[19]]
combine_pred(newdata[newdata$set == 0,], L)

if (F) {
    # mr data
    preds = sapply(L, function(model) {
        pred = predict(model, newdata = newdata, type = "lp")
    })
    preds = as.data.frame(preds)
    colnames(preds) = c('mr_fold1', 'mr_fold2', 'mr_fold3', 'mr_fold4')
    rownames(preds) = newdata$name

    # clinic data
    clinic.preds = sapply(L, function(model) {
        pred = predict(model, newdata = newdata, type = "lp")
    })
    colnames(clinic.preds) = c('cli_fold1', 'cli_fold2', 'cli_fold3', 'cli_fold4')
    rownames(clinic.preds) = newdata$name
    for (col in colnames(clinic.preds)) {
        preds[,col] = ifelse(rownames(preds) %in% rownames(clinic.preds), clinic.preds[match(rownames(preds), rownames(clinic.preds)), col], NA)
    }

    deep = c(
        -0.02898905,  0.1292885 , -0.1320485 ,  0.0190624 ,  0.2206999 ,
        0.05161246,  0.05798227,  0.14294168,  0.13917539,  0.20027529,
        0.21048102,  0.11443229,  0.305116  ,  0.05373815,  0.0330108 ,
        0.06395182,  0.0516629 ,  0.17723022, -0.03947562, -0.10685159,
        0.11662177,  0.01420938,  0.08733527,  0.18295585,  0.22989021,
        0.05862349,  0.14375281,  0.04727061, -0.04779226,  0.12051404,
        0.1217624 ,  0.13983373,  0.07681957,  0.1021561 ,  0.0485805 ,
        0.23270428,  0.08740444,  0.11307081,  0.08562877,  0.02192201,
        0.10035942,  0.02732991,  0.01492278,  0.23744978,  0.1029411 ,
        0.01723914, -0.03577077,  0.18610285,  0.02136759,  0.07142243,
        0.3049452 ,  0.01409854, -0.0122322 ,  0.17929362,  0.06286237,
        -0.07825693, -0.09075248,  0.00772203,  0.15160061,  0.08312748,
        0.04605741,  0.05698361,  0.01683023,  0.09370644,  0.1505618 ,
        0.14300813,  0.14541636,  0.12497155,  0.22723673,  0.12412071,
        0.19362962,  0.07452842,  0.07285443,  0.04068792,  0.1476985 ,
        0.00731126,  0.27930304,  0.01805751, -0.01898891, -0.10380702,
        0.16159646,  0.09229032,  0.22712572, -0.00509511,  0.08644661,
        0.28520343,  0.0614658 ,  0.13266943,  0.20343123, -0.04204567,
        -0.01601395,  0.11731911, -0.0426845 ,  0.04255704,  0.01074051,
        0.08567341,  0.04958091,  0.16283928,  0.0720663 , -0.04042432,
        0.14671472,  0.09145056,  0.18629132,  0.01279549,  0.25326177,
        0.13143866, -0.0138239 ,  0.13551915, -0.05002581,  0.02882713,
        -0.03124199,  0.17793004,  0.1534499 ,  0.15872175,  0.213689  ,
        0.02025224,  0.00347952, -0.00442515,  0.2691839 , -0.02758418,
        -0.05877576,  0.06730221,  0.09354333, -0.05192444,  0.11794921,
        0.15109257,  0.11756733, -0.12787367, -0.06864613, -0.00779666,
        0.07217324,  0.12857197,  0.10810547,  0.28268048,  0.01090285,
        -0.07657885,  0.04064992,  0.41972116,  0.08760549, -0.02887081,
        0.18250999,  0.03355498,  0.09787277,  0.01964291, -0.05550218,
        0.23859447,  0.14470427,  0.15245022,  0.21261892,  0.12887798,
        0.19980185, -0.01104072,  0.0090052 ,  0.15360545,  0.05700964,
        0.01036449, -0.01198922,  0.21742046,  0.17691323,  0.03378471,
        0.10185765,  0.02836291,  0.2227042 ,  0.08985304, -0.07072676,
        0.00537774,  0.02497649,  0.07771533,  0.22510427,  0.10807826,
        0.10545082,  0.13866763,  0.18048428, -0.0258197 ,  0.12574486,
        0.18784368,  0.13415192,  0.16444269,  0.17951539,  0.0962243 ,
        0.24064715,  0.18957877, -0.01095681, -0.04635367,  0.11971303,
        -0.08836257,  0.08163248,  0.31167042, -0.07346828,  0.14839141,
        0.1310746 ,  0.08188711, -0.04377233, -0.02433155,  0.16302426,
        0.07217969,  0.16401906, -0.08124782,  0.11568806,  0.15750039,
        0.0082744 ,  0.02910628,  0.04231551,  0.06614181,  0.22310875,
        0.01797072,  0.05073438,  0.21446165,  0.03177476,  0.01634365,
        0.17802395,  0.11059649,  0.08008953, -0.05989357,  0.04350982,
        0.13099693,  0.23056561,  0.12422559,  0.06289749,  0.2984874
    )
    deeppat = c(
        1602186, 1415838, 1503058, 1403521, 1507884, 1602751, 1522757,
        1401438, 1422007, 1608839, 1505470, 1700618, 1700457, 1400860,
        1508005, 1508163, 1502675, 1516928, 1501049, 1506707, 1700572,
        1602586, 1508191, 1522923, 1410915, 1511058, 2154014, 1615495,
        1612055, 1301630, 1504545, 1602137, 1405035, 1507217, 1509570,
        1404316, 1505902, 1510941, 1410116, 1700342, 1520851, 1613971,
        1503718, 1519706, 1405036, 1414189, 1603444, 1507472, 1408401,
        1600296, 1520263, 1608980, 1700177,  153461, 1702746, 1507248,
        1701663, 1706442, 1518062, 1501820, 1501611, 1512396, 1512377,
        1510442, 1704471, 1605606, 1503566, 1505796, 1516078, 1401150,
        1615043, 1701668, 1608463, 1509038, 1600870, 1501050, 1401317,
        1602296, 1614914, 1409660, 1610465, 1405152, 1805044, 1406489,
        1401942, 1401920, 1613361, 1504186, 1315715, 1509767, 1503005,
        1604547, 1605686, 1513218, 1615309, 1404637, 1701138, 1611422,
        1510827, 1506777, 1504406, 1614456, 1423521, 1702589, 1601181,
        1503175, 1605385, 1406199, 1407127, 1701475, 1604103, 1406407,
        1403602, 1604044, 1405481, 1315312, 1500694, 1609874, 1501010,
        1601789, 1605530, 1703032, 1508362, 1607174, 1605283, 1501417,
        1701811, 1608527, 1406322, 1502357, 1510358, 1601095, 1403311,
        1507859, 1501415, 1601350, 1700462, 1511384, 1607688, 1604176,
        1408018, 1410317, 1407606, 1607972, 1600294, 1510392, 1510123,
        1508384, 1422841, 1404116, 1402355, 1501416, 1702593, 1604426,
        1507284, 1605918, 1505414, 1419091,  140633, 1513602, 1402299,
        1604846, 1300134, 1410575, 1505679, 2134639, 1601239, 1509612,
        1400974, 1607236, 1404315, 1523839,  152216, 1504357, 1403144,
        1613107, 1600991, 1402805, 1509056, 1604060, 1508466, 1515420,
        1614913, 1610256, 1403257, 1409419, 1402909, 1615914, 1510039,
        1404924, 1316664, 1411030, 1602620, 1408272, 1512101, 1512384,
        1511382, 1602926, 1405541, 1407348, 1607437, 1510261, 1313783,
        1602238, 1501684, 1403098, 1614786, 1602453, 1508175, 1602748,
        154320, 1500695, 1404309, 1702583, 1610180, 1603591, 1301415,
        1505057, 1510385, 1306115
    )

    deepname = labeldata$name[match(deeppat, labeldata$number)]
    preds$deep = deep[match(rownames(preds), deepname)]
    preds$name = rownames(preds)
    preds$event = newdata$event[match(newdata$name, preds$name)]
    preds$time = newdata$time[match(newdata$name, preds$name)]

    write.csv(preds, "/home/tongxueqing/zhao/ImageProcessing/combine_model/_data/preds.csv")
}
